Refactor: OAuth 2.0 사용자 클래스를 model 패키지 안으로 이동 #13
