+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/auth/model/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package com.nextdoor.nextdoor.domain.auth;
+package com.nextdoor.nextdoor.domain.auth.model;
 
 import java.util.Collection;
 import java.util.Collections;
